Move UplynkSsaiDeserializer to internal
